"On Demand" mode.
Adds an `on_demand` config option that prevents the supervisor from
starting up initially. Instead, it spawns an Agent process to store the
configuration for us, and dynamically launches the supervisor (unnamed)
as needed, and stops it after use.

This should help reducing the "zombie process" pain in development when
stopping a server / iex shell with Ctrl+C.

Addresses #56.