Include inspected protocol in timeout error
This patch adds an `Inspect` implementation for `ChromicPDF.Protocol`
that redacts any potentially sensitive data, leaving merely ChromicPDF
internals and options. Next, the inspected protocol is included in the
timeout error in `Channel.run_protocol/3`, as this is the error a lot of
people run into in various situations: boot time, teardown, etc. See
issues #160, #171 for instance.