[#219] Handle DOWN in terminate_worker/3
This patch handles the `DOWN` reason in the `terminate_worker/3`
callback, which we previously thought "would not happen", but a user
reported it as "does happen indeed". Not 100% sure how or why it
happens, but we don't expect anything to break if we handle it.

Fixes [#219](https://github.com/bitcrowd/chromic_pdf/issues/219)