[#290] Fix missing footers
Fixes [[#290]](https://github.com/bitcrowd/chromic_pdf/issues/290)

Since Chrome 117 custom footers would not be displayed anymore when
configured via `ChromicPDF.Template`. Reason for this was that our
stylesheet was prepended to both the content markup as well as the
header and footer templates. The `@page` rule contains the page margins
which used to be ignored in the header and footer sections, but are not
applied and lead to the footer being pushes out of its content box.
Fixed by splitting `ChromicPDF.Template.styles/1` into `page_styles/1`
and `header_footer_styles/1` and removing the `@page` directive from the
header and footer styles.

Besides, this patch removes the `zoom` directive from the header and
footer templates for Chrome >= v120, as they have removed the default
4/3 scale-up.