Remove Alpine 3.16 from test matrix (#315)
This removes our Alpine 3.16-based test image as it was heavily flaky (failing with timeouts
in about every other test run) and we have no idea how to fix it. On the other hand, we have
run ChromicPDF on Alpine 3.16 for years and it worked perfectly fine. Additionally, the
situation seems to have improved with more recent Chromiums a lot, so that the other test
images do not have timeout issues anymore. Alpine 3.16 is also quite old by now, so I conclude
we're ok to just drop it, in favor of having green test runs again some day.