Support custom protocols
This patch makes the API more versatile by allowing to run custom protocols.

- add :protocol option to `print_to_pdf/2` & `capture_screenshot/2`
- refactor API & options handling around `chrome_export/2`
- add new top-level but hidden `run_protocol/2` API

For the time being, these features are not advertised in the documentation. The
`ChromicPDF.Protocol` machinery can be quite tricky and the poor DSL in
`ChromicPDF.ProtocolMacros` is subject to change. Nonetheless, this could be a way
to avoid adding further options to change and extend the default protocols' behaviour.

Relates to #319